API Testing: Retrieved a single tweet and stored its JSON data locally.
Utilized the API to retrieve JSON data for a single tweet. That JSON
string was then stored as a single line to a local .txt file. Then that .txt file was read and select parameters of the tweet were extracted
and stored to a local dataframe (i.e., local to the jupyter notebook
workspace). Next steps include repeating this behavior for all tweets,
thereby creating a much larger .txt file, along with a pandas dataframe
with useful tweet information.