API Working: Retrieved and saved JSON info for all tweets of interest.
The API is working now. JSON info for all tweets of interest has been
retrieved and saved to 'tweet_json.txt'. Additionally, desired metadata
was extraced from 'tweet_json.txt' and stored to a dataframe in the
Jupyter Notebook workspace. This likely concludes the 'Gather' phase
of this Data Wrangling project. Next steps involve assessing, cleaning,
and visualizing the data.