Updated Doc and Assessment: Edited README.md, continued data assessment
README.md now has notes on how to view HTML versions of a Jupyter
Notebook on GitHub. Once the analysis is completed, and HTML version
will be uploaded and this will actually become relevant. Additionally,
the various dataframes were assessed for Accuracy, Validity, and
Completeness. The assessment is ongoing, but likely close to being
finished.