Cleanup: Cleaned the .ipynb file and added master .csv files
- 'Temporary' dataframe variables were removed from the .ipynb file,
then the kernel was restarted, and all cells were ran. This is likely
to be the final version of the .ipynb file, unless the analysis is
revisited at a future point.
- Added a .csv file for the two 'master' dataframes (i.e., the final
dataframes, after they have been cleaned and have had some light
feature engineering performed on them).
- Next, and possibly final, step is to add an 'external' report, which
will take the form of a blog post.